Working render loop with simple vkCmdClearColorImage. We also fixed the following issue:
It turns out that it isn't sufficient to have frames in flight number of present wait semaphores. You need as many present wait semaphores as you do have swapchain images because the only way to guarantee that the semaphore being waited by vkQueuePresentKHR has signaled and presentation has happened is if you acquire the same image in a future frame. However, if you don't and try to use the same semaphore, there's no guarantee that it isn't being waited on.