mesh model-view transformations are now served up via a per-frame transform buffer.
transforms are computed after the fence to ensure we're not writing to the shader storage buffer whilst it is being read by a device thread.

next up is some much needed cleanup. we need to think of a way to tie all these resources we manage nicely on the host.