Texture indices are now stored in materials.
Trying out a coding scheme where variables are prefixed indicating whether they are members of a structure and their type. For example, m_v3Position.

Lighting is now computed in world space.

GlobalData is now updated each frame.

General cleanup of the data that is passed to the GPU. We'll be introducing a physically-based BRDF and so now was a good time to trim this down.